Use new PyBossa.JS pre-loading tasks pattern
This commit uses the new PyBossa.JS pre-loading tasks pattern that allows you
to load a second task in the background while the user is solving the first one.

This allows to improve the UX as loading the next task is almost instantly
(unless the user answers faster than downloading the next image).

Additionally, a new layout has been added that allows to only load the disqus
comments on-demand.